+++ translations/it.php
@@ -1,0 +1,11 @@
+<?php
+
+global $_MODULE;
+$_MODULE = array();
+$_MODULE['<{cleanshare}prestashop>cleanshare_5a95a425f74314a96f13a2f136992178'] = 'Condividi';
+$_MODULE['<{cleanshare}prestashop>cleanshare_1fe381cb8750248dabf09487bc01f46d'] = 'Condividi';
+$_MODULE['<{cleanshare}prestashop>cleanshare_page_0a3e04ecafd1b5a9a500783d4fbdb184'] = 'Copia fallita. Url: %s';
+$_MODULE['<{cleanshare}prestashop>cleanshare_page_fdac44da058fb4d8f9da57322a99c26b'] = 'Copia non supportata. Url: %s';
+$_MODULE['<{cleanshare}prestashop>cleanshare_page_43031118a545608417532d43c55fb9bd'] = 'URL copiato negli appunti';
+$_MODULE['<{cleanshare}prestashop>cleanshare_float_4da480e6df3ee3a2a3f8ba1b01038b55'] = 'URL copiato negli appunti';
+$_MODULE['<{cleanshare}prestashop>cleanshare_button_4da480e6df3ee3a2a3f8ba1b01038b55'] = 'URL copiato negli appunti';